@@ -35,6 +35,7 @@
     "@sabinmarcu/config": "workspace:*",
     "@sabinmarcu/debug": "workspace:*",
     "@sabinmarcu/eslint-config": "workspace:*",
+    "@sabinmarcu/jotai-storage-adapter-querystring": "workspace:*",
     "@sabinmarcu/observable": "workspace:*",
     "@sabinmarcu/types": "workspace:*",
     "@sabinmarcu/utils-fs": "workspace:*",
@@ -42,7 +43,7 @@
     "@sabinmarcu/utils-primitives": "workspace:*",
     "@sabinmarcu/utils-repo": "workspace:*",
     "@sabinmarcu/utils-test": "workspace:*",
-    "typescript": "5.3.2"
+    "typescript": "5.5.4"
   },
   "browserslist": {
     "production": [
