@@ -12,7 +12,11 @@
     "@tscmono/cli": "^0.2.2",
     "@tscmono/preset-default": "^0.4.6",
     "babel-plugin-module-rewrite": "^0.2.0",
-    "eslint": "^8.30.0",
+    "eslint": "^9.8.0",
+    "eslint-plugin-expect-type": "^0.4.0",
+    "eslint-plugin-jsx-a11y": "^6.9.0",
+    "eslint-plugin-react": "^7.35.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "glob": "^8.0.3",
     "glob-tester-cli": "^0.1.0",
     "husky": "^8.0.2",
@@ -20,9 +24,10 @@
     "minimist": "^1.2.7",
     "mkdist": "^1.0.0",
     "moize": "^6.1.4",
-    "typescript": "5.3.2",
+    "typescript": "5.5.4",
+    "typescript-eslint": "^8.0.0",
     "unbuild": "^1.0.0",
-    "yaml": "^2.2.0"
+    "yaml": "^2.5.0"
   },
   "scripts": {
     "coverage": "jest --coverage",
@@ -34,7 +39,7 @@
   },
   "devDependencies": {
     "@commitlint/cz-commitlint": "^17.3.0",
-    "@types/eslint": "^8.4.10",
+    "@types/eslint": "^9.6.0",
     "@types/glob": "^8",
     "@types/inquirer": "^8",
     "@types/jest": "^29.2.4",
