@@ -1,31 +1,37 @@
 import type {
   Config,
-} from './types.js';
+} from './types';
 
-let extension: string = '';
-try {
-  extension = __filename.split('.').pop()!;
-} catch {
-  // @ts-ignore
-  extension = import.meta.url.split('.').pop()!;
-} finally {
-  extension = `.${extension}`;
+import parserConfig from './configs/parser';
+import jsConfig from './configs/js';
+import tsConfig from './configs/ts';
+import rootConfigsConfig from './configs/config.root';
+import jestConfigsConfig from './configs/config.jest';
+import moduleConfig from './configs/module';
+import storybookConfig from './configs/storybook';
+import unicornConfig from './configs/unicorn';
+import typeTestingConfig from './configs/typeTesting';
+import canonicalConfig from './configs/canonical';
+
+const hasParser = parserConfig.length > 0;
+
+if (!hasParser) {
+  console.error('No parser installed, will not be loading config!');
 }
 
-const config = {
-  extends: [
-    './configs/unicorn',
-    './configs/js',
-    './configs/jsx',
-    './configs/ts',
-    './configs/tsx',
-    './configs/ts.expect',
-    './configs/module',
-    './configs/stories',
-    './configs/jest',
-    './configs/config',
-    './configs/overrides',
-  ].map((path) => require.resolve(path + extension)),
-} satisfies Config;
+const config: Config[] = hasParser
+  ? [
+    ...parserConfig,
+    ...jsConfig,
+    ...tsConfig,
+    ...rootConfigsConfig,
+    ...jestConfigsConfig,
+    ...moduleConfig,
+    ...storybookConfig,
+    ...unicornConfig,
+    ...typeTestingConfig,
+    ...canonicalConfig,
+  ] as const satisfies Config[]
+  : [] satisfies Config[];
 
-module.exports = config;
+export default config;
