@@ -1,6 +1,6 @@
 import type {
   Config,
-} from './types.js';
+} from './types';
 
 type _Test = Config;
-//    ^? type _Test = Linter.Config<Linter.RulesRecord, Linter.RulesRecord>
+//    ^? type _Test = Linter.Config<Linter.RulesRecord>
