@@ -1,4 +1,4 @@
-/* eslint-disable import/no-extraneous-dependencies */
+ 
 
 const { pathToFileURL } = await import('node:url');
 const { smart } = await import('@babel/template');
