@@ -3,7 +3,7 @@ import {
 } from '../../../.config/jest/jest.config.project.mjs';
 
 export default generateFromPath(
-  (await import('url')).fileURLToPath(
+  (await import('node:url')).fileURLToPath(
     new URL('.', import.meta.url),
   ),
 );
